[ADD] context graphql module - modulize context function
